# EDSC-4344: Refresh order status page faster while order is still creating