--- a/static/src/js/components/OrderStatus/OrderStatusItem.jsx
+++ b/static/src/js/components/OrderStatus/OrderStatusItem.jsx
@@ -11,8 +11,7 @@ import {
 import { FaQuestionCircle } from 'react-icons/fa'
 import moment from 'moment'
 
-import { getClientId } from '../../../../../sharedUtils/getClientId'
-import { getApplicationConfig, getEnvironmentConfig } from '../../../../../sharedUtils/config'
+import { getApplicationConfig } from '../../../../../sharedUtils/config'
 import {
   getStateFromOrderStatus,
   aggregatedOrderStatus,
@@ -30,63 +29,13 @@ import S3LinksPanel from './OrderStatusItem/S3LinksPanel'
 import DownloadScriptPanel from './OrderStatusItem/DownloadScriptPanel'
 import BrowseLinksPanel from './OrderStatusItem/BrowseLinksPanel'
 import getOrderInfoAndProgress from './getOrderInfoAndProgress'
+import buildEddLink from './buildEddLink'
 
 import ProgressRing from '../ProgressRing/ProgressRing'
 
 import './OrderStatusItem.scss'
 
-const buildEddLink = ({
-  authToken,
-  collection,
-  downloadUrls,
-  earthdataEnvironment,
-  linkType
-}) => {
-  const {
-    collection_metadata: collectionMetadata,
-    orders = [],
-    retrieval_collection_id: retrievalCollectionId
-  } = collection
-
-  const orderStatus = aggregatedOrderStatus(orders)
-
-  const [firstOrder = {}] = orders
-  const {
-    type = ''
-  } = firstOrder
-
-  // If the order is Harmony and is still running or has no files, don't show the EDD link
-  const isDone = !['creating', 'in progress'].includes(orderStatus)
-  const notDoneOrEmpty = !isDone || downloadUrls.length === 0
-  if (type.toLowerCase() === 'harmony' && notDoneOrEmpty) {
-    return null
-  }
-
-  const {
-    conceptId,
-    shortName,
-    versionId
-  } = collectionMetadata
-
-  let downloadId = conceptId
-  if (shortName) downloadId = `${shortName}_${versionId}`
-
-  // Build the `getLinks` URL to tell EDD where to find the download links
-  const { apiHost, edscHost } = getEnvironmentConfig()
-  const getLinksUrl = `${apiHost}/granule_links?id=${retrievalCollectionId}&flattenLinks=true&linkTypes=${linkType}&ee=${earthdataEnvironment}`
-
-  // Build the authUrl to tell EDD how to authenticate the user
-  const authReturnUrl = 'earthdata-download://authCallback'
-  const authUrl = `${apiHost}/login?ee=${earthdataEnvironment}&eddRedirect=${encodeURIComponent(authReturnUrl)}`
-
-  // Build the eulaRedirectUrl to tell EDD how to get back after the user accepts a EULA
-  const eulaCallback = 'earthdata-download://eulaCallback'
-  const eulaRedirectUrl = `${edscHost}/auth_callback?eddRedirect=${encodeURIComponent(eulaCallback)}`
-
-  const link = `earthdata-download://startDownload?getLinks=${encodeURIComponent(getLinksUrl)}&downloadId=${downloadId}&clientId=${getClientId().client}&token=Bearer ${authToken}&authUrl=${encodeURIComponent(authUrl)}&eulaRedirectUrl=${encodeURIComponent(eulaRedirectUrl)}`
-
-  return link
-}
+const { orderStatusRefreshTime, orderStatusRefreshTimeCreating } = getApplicationConfig()
 
 let intervalId = null
 
@@ -191,8 +140,6 @@ export const OrderStatusItem = ({
   // Handles fetching the retrieval collection to get the order info
   useEffect(() => {
     const { type: accessMethodType } = accessMethod
-
-    const { orderStatusRefreshTime, orderStatusRefreshTimeCreating } = getApplicationConfig()
 
     let refreshInterval = orderStatusRefreshTime