--- a/static/src/js/components/Well/Well.scss
+++ b/static/src/js/components/Well/Well.scss
@@ -13,6 +13,7 @@
   &__heading {
     font-weight: 700;
     font-size: 1.5rem;
+    letter-spacing: -0.015rem;
   }
 
   &__introduction {