# Overview

### What is the feature?

Refresh order status page faster while order is still creating

### What is the Solution?

While the order is in the `creating` state, fetch the order status every 5s (configurable).
Refactor the OrderStatusItem component to be a functional component with Testing Library tests
Adds the last updated day/time to the order status page. This is the last time our backend checked the status of the order (every 60s)

### What areas of the application does this impact?

Order status page

# Testing

Locally run EDSC with `npm run start:optionals`
Submit an order, watch the network tab to see how often the `retrieval_collections/{id}` endpoint is called. While the order is `creating` it will be every 5s. After it moves to In Progress it will refresh every 60s 

This [Harmony job](http://localhost:8080/projects?p=C1595422627-ASF!C1595422627-ASF&pg[1][a]=2736530662!2736533197!2736534372!2736533270!ASF&pg[1][v]=t&pg[1][gsk]=-start_date&pg[1][m]=harmony0&pg[1][uv]=V2666154475-ASF&pg[1][cd]=f&pg[1][ets]=t&pg[1][ess]=f&q=C1595422627-ASF&sb[0]=-77.16797%2C-43.06761%2C-63.84375%2C-31.25757&qt=2023-05-11T18%3A55%3A06.122Z%2C2023-06-26T09%3A13%3A53.645Z&tl=1693663763!3!!&lat=-40.253225239533435&long=-84.515625&zoom=4) works well 

# Checklist

- [x] I have added automated tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings